User Invites
Got user invites working with help from the Yii2 Advanced Template.

Admins with userCreate can now invite users and the new user must click on a link in their email to verify their account and set a new password.

Just need to work on correcting the link that is sent out in the emails